requirements.txt
kaggle
pandas
scikit-learn
pydicom
Pillow